reporter: Merge `NoticeReporter` and `AbstractNoticeReporter`
After the removal of `NoticeOnlyDeclaredReporter` in the previous commit
there is not reason to keep `NoticeReporter`and `AbstractNoticeReporter` as
separate classes.
